最後の管理者は削除できない

+++ usecase/administrator/delete_administrator.go
@@ -2,6 +2,7 @@ package administrator
 
 import (
 	"context"
+	"errors"
 
 	"github.com/Cityboypenguin/SPACE-server/repository"
 )
@@ -23,5 +24,12 @@ func NewDeleteAdministratorUseCase(adminRepo repository.AdministratorRepository)
 }
 
 func (uc *DeleteAdministratorInteractor) Execute(ctx context.Context, id int64) (bool, error) {
+	admins, err := uc.adminRepo.ListAdministrators(ctx)
+	if err != nil {
+		return false, err
+	}
+	if len(admins) <= 1 {
+		return false, errors.New("最後の管理者は削除できません")
+	}
 	return uc.adminRepo.DeleteAdministrator(ctx, id)
 }
